+++ starter/src/tools/gridtools2.ts
@@ -1,3 +1,4 @@
+import writeXlsxFile from "write-excel-file"
 import { fetchData } from "../api/fetchData"
 import AppContext from "../AppContext"
 import { IDescrFilter, } from '../types'
@@ -102,8 +103,33 @@ const postGrid = async (rows: any, data: any, descrType: string) => {
    return `Number of affected rows ${result||-1}`
 }
 
-const toExcel = async (rows: any) => {
+const toExcel = async (cols: any, rows: any) => {
     console.log('toExcel', rows)
+
+    const exportCols = cols.map((col: any) => {
+        return {value: col.name, fontWeight: 'bold'}
+    })
+
+    const exportRows = rows.map((row: any) => {
+        const exportRow: any[] = []
+        cols.map((col: any) => {
+            exportRow.push({ value: row[col.key] })
+        })
+        return exportRow
+    })
+    exportRows.unshift(exportCols)
+
+    try {
+        const xlsxBlob = await writeXlsxFile(exportRows, {
+            // schema: columnsDefinition,
+            // https://gitlab.com/catamphetamine/write-excel-file#browser
+            // when no fileName is given results in blob
+            fileName: `temp_data.xlsx`,
+        });
+        console.log("xlsxBlob", xlsxBlob);
+    } catch (error) {
+        console.log(error);
+    }
 }
 
 export { getGridCols, getGridRows, postGrid, toExcel, }